Post 수정, 삭제 기능 구현 (issue: #3) (#7)
* modify : Post List 삭제 안된 List만 받아오게 수정

* fix : Board별로 저장되지 않고 항상 같은 Board에만 저장되는 Bug Fix

* feat : Post Delete 기능 구현 및 Board Delete 로직 수정

* feat : Post Edit 기능 구현

* modify : Delete Api ID로 삭제하게 변경

+++ frontend/simple_board/lib/model/reply_entity.dart
@@ -8,8 +8,8 @@ part 'reply_entity.g.dart';
 
 @JsonSerializable()
 class ReplyEntity implements EntityBase {
-    double id;
-    double postId;
+    int id;
+    int postId;
     String userName;
     @override
     String password;
